Refactored bot application code into smaller, more related parts
Now the 'bot' directory, where the bot code resides, contains 3 source codes:
	- bot.py: Entrance point for the bot. Deals with various initializations
	- bot_handlers.py: defines a class that stores the bot handler functions (usually just callback functions)
		that are executed in accordance with the description on the 'bot.py' file.
	- redis_operations: defines a class that stores the connection to the redis database (for storing information
		about the users and their relationship with Spotify) and functions that mostly work with the database, serving
		to facilitate DB interection from the handles

Another change was the reorganization of the necessary "global variables" (the variables previously all defined on .env files) to maintain
variables that are secure to mantain the on the 'config.yaml' file and private variables, as API tokens, on the .env files. For this reason,
they are not included with this commit. The variables that need to be included in a .env for the bot as 'TELEGRAM_TOKEN' (the Telegram Bot token),
'SPOTIFY_CLIENT_ID' (the id of the Spotify API App) and 'SPOTIFY_CLIENT_SECRECT' (the secret id of the Spotify API App). For the webserver, only
the necessary Flask environment variables.

Note: redis server setup not included with this commit. Currently, Redis server is only running local
Note: Docker Compose services are still not ready. As of now, using each process on multiple terminals