Separation between bot handlers and their implementation (operations using Spotify endpoits were separeted)
This commit represents the start of the implementation of our main feature: the population of a Spotify Playlist with
recommended songs based on a questionary. First, acess scope to the Spotify account was expanded. Second, there has been made
a separation between the logic of the handles (and its messages) from the implementation of functions that acess various Spotify
Endpoints. In the future, this file will store functions capable of getting the Spotify User ID, create a private Spotify playlist,
and so on...